requirements.txt
PyMuPDF==1.25.5
openai==1.72.0
langdetect==1.0.9
click==8.1.7
tqdm==4.67.1
pytest==8.3.5
flask==3.1.0
python-dotenv==1.0.1
